Refactor repository/query patterns (#1695)
